+++ tests/testthat.R
@@ -9,4 +9,21 @@
 library(testthat)
 library(Giotto)
 
+# additional needed packages
+suppressWarnings({
+  if(!require(remotes)){
+    install.packages('R.utils', repos = 'http://cran.us.r-project.org')
+    install.packages('remotes', repos = 'http://cran.us.r-project.org')
+  }
+
+  if(!require(GiottoData)){
+    remotes::install_github('drieslab/GiottoData')
+  }
+})
+
+# install giotto environment
+if (!checkGiottoEnvironment()) {
+  installGiottoEnvironment()
+}
+
 test_check("Giotto")