+++ tests/testthat/test_accessors.R
@@ -1,20 +1,5 @@
 
-
-suppressWarnings({
-  if(!require(remotes)){
-    install.packages('R.utils', repos = 'http://cran.us.r-project.org')
-    install.packages('remotes', repos = 'http://cran.us.r-project.org')
-  }
-
-  if(!require(GiottoData)){
-    remotes::install_github('drieslab/GiottoData')
-  }
-
-  require(testthat)
-})
-
-
-library(Giotto)
+require(testthat)
 
 # Load subobjects
 ex = GiottoData::loadSubObjectMini('exprObj')