+++ tests/testthat/test_createObject.R
@@ -1,18 +1,5 @@
 
-
-suppressWarnings({
-  if(!require(remotes)){
-    install.packages('R.utils', repos = 'http://cran.us.r-project.org')
-    install.packages('remotes', repos = 'http://cran.us.r-project.org')
-  }
-
-  if(!require(GiottoData)){
-    remotes::install_github('drieslab/GiottoData')
-  }
-
-  require(testthat)
-})
-
+require(testthat)
 
 # Load subobjects
 ex = GiottoData::loadSubObjectMini('exprObj')