+++ tests/testthat/test_dataImports.R
@@ -3,19 +3,7 @@
 #   installGiottoEnvironment()
 # }
 
-suppressWarnings({
-  if(!require(remotes)){
-    install.packages('R.utils', repos = 'http://cran.us.r-project.org')
-    install.packages('remotes', repos = 'http://cran.us.r-project.org')
-  }
-
-  if(!require(GiottoData)){
-    remotes::install_github('drieslab/GiottoData')
-  }
-
-  require(testthat)
-})
-
+require(testthat)
 
 
 ### TESTS FOR DATA IMPORT FUNCTIONS