+++ tests/testthat/test_gobject.R
@@ -1,17 +1,6 @@
 
 
-suppressWarnings({
-  if(!require(remotes)){
-    install.packages('R.utils', repos = 'http://cran.us.r-project.org')
-    install.packages('remotes', repos = 'http://cran.us.r-project.org')
-  }
-
-  if(!require(GiottoData)){
-    remotes::install_github('drieslab/GiottoData')
-  }
-
-  require(testthat)
-})
+require(testhat)
 
 
 # Load subobjects