+++ NEWS.md
@@ -3,7 +3,7 @@
 # Giotto Suite 3.3.1 (Release TBD)
 
 ## Breaking Changes
-
+- Change `checkGiottoEnvironment()`. Downgrade from error to message and return FALSE when a provided directory does not exist
 
 ## Added
 - New file `poly_influence.R`